Change the location of the git-bridge data directory
The default location of the data directory inside the git-bridge
container is /tmp/wlgb, but the data it contains should not be
considered temporary. In order to avoid any misunderstanding that would
lead an administrator to delete the contents of the /tmp directory, we
move the data directory to /data/git-bridge.