Show container logs via bin/logs
Most of our service logs are found inside the sharelatex container in
/var/sharelatex/logs, but some service logs (for mongo, redis,
git-bridge) are found in the container logs. The bin/logs script can now
show both types of logs.

In addition, the -n option, which controls how many lines of logs are
output now accepts the string "all" to make it easy to show all
available logs for a given service.

This is meant to be used to easily extract logs with:

    bin/logs -n all service > service.log