Properly kill logging processes when using Compose v2
When killing a `docker compose exec` process with Compose v2, the kill
signal somehow doesn't propagate to the process running inside the
container. As a workaround, we collect the PIDs from inside the
container when we start a process in the background and kill the
processes from inside the container at the end.